fix: anchor background tasks + drop deprecated typing usages
Static audit (ruff F/E/W/B/RUF + pyflakes + vulture) surfaced four
issues; pyflakes and ruff are now clean.

- bot.py: store every asyncio.create_task() return in a module-level
  _BG_TASKS set with a done_callback to discard. The auto-delete TTL
  task at the reply path was fire-and-forget with no strong ref - per
  asyncio docs the task could be garbage-collected mid-sleep, silently
  killing the auto-delete. Also covers any future fire-and-forget work.
- bot.py: type _STATUS_PAGES with collections.abc.Callable instead of
  the builtin 'callable' (which is not a valid type), removing the
  '# type: ignore[type-arg]' workaround.
- analytics.py: 'from typing import Iterator' -> 'collections.abc'
  (typing.Iterator deprecated since Python 3.9, removed in 3.13).
- analytics.py: drop redundant int(round(...)) cast; round() already
  returns int when ndigits is omitted.

+++ analytics.py
@@ -14,9 +14,9 @@ import os
 import sqlite3
 import threading
 import time
+from collections.abc import Iterator
 from contextlib import contextmanager
 from pathlib import Path
-from typing import Iterator
 
 logger = logging.getLogger(__name__)
 
@@ -127,7 +127,7 @@ def _percentile(values: list[int], pct: float) -> int:
     if not values:
         return 0
     s = sorted(values)
-    k = max(0, min(len(s) - 1, int(round((pct / 100.0) * (len(s) - 1)))))
+    k = max(0, min(len(s) - 1, round((pct / 100.0) * (len(s) - 1))))
     return s[k]
